implement slots and dependencies
